+++ integration-tests/api/__tests__/store/gift-cards.js
@@ -1,9 +1,9 @@
 const path = require("path")
 const { Region, GiftCard } = require("@medusajs/medusa")
 
-const setupServer = require("../../../helpers/setup-server")
-const { useApi } = require("../../../helpers/use-api")
-const { initDb, useDb } = require("../../../helpers/use-db")
+const setupServer = require("../../../environment-helpers/setup-server")
+const { useApi } = require("../../../environment-helpers/use-api")
+const { initDb, useDb } = require("../../../environment-helpers/use-db")
 
 jest.setTimeout(30000)