+++ packages/product/src/types/index.ts
@@ -1,17 +1,4 @@
-import { Constructor, DAL, IEventBusService } from "@medusajs/types"
-
-export type ProductServiceInitializeOptions = {
-  database: {
-    clientUrl: string
-    schema?: string
-    driverOptions?: Record<string, unknown>
-  }
-}
-
-export type ProductServiceInitializeCustomDataLayerOptions = {
-  manager?: any
-  repositories?: { [key: string]: Constructor<DAL.RepositoryService> }
-}
+import { IEventBusService } from "@medusajs/types"
 
 export type InitializeModuleInjectableDependencies = {
   eventBusService?: IEventBusService