+++ packages/types/src/product/common.ts
@@ -34,6 +34,7 @@ export interface ProductDTO {
   tags: ProductTagDTO[]
   variants: ProductVariantDTO[]
   options: ProductOptionDTO[]
+  images: ProductImageDTO[]
   discountable?: boolean
   external_id?: string | null
   created_at?: string | Date
@@ -86,7 +87,7 @@ export interface ProductTagDTO {
   id: string
   value: string
   metadata?: Record<string, unknown> | null
-  products: ProductDTO[]
+  products?: ProductDTO[]
 }
 
 export interface ProductCollectionDTO {
@@ -95,6 +96,7 @@ export interface ProductCollectionDTO {
   handle: string
   metadata?: Record<string, unknown> | null
   deleted_at?: string | Date
+  products?: ProductDTO[]
 }
 
 export interface ProductTypeDTO {
@@ -109,6 +111,13 @@ export interface ProductOptionDTO {
   title: string
   product: ProductDTO
   values: ProductOptionValueDTO[]
+  metadata?: Record<string, unknown> | null
+  deleted_at?: string | Date
+}
+
+export interface ProductImageDTO {
+  id: string
+  url: string
   metadata?: Record<string, unknown> | null
   deleted_at?: string | Date
 }
@@ -163,4 +172,127 @@ export interface FilterableProductCategoryProps
   is_active?: boolean
   is_internal?: boolean
   include_descendants_tree?: boolean
+}
+
+/**
+ * Write DTO (module API input)
+ */
+
+export interface CreateProductTypeDTO {
+  id?: string
+  value: string
+}
+
+export interface CreateProductTagDTO {
+  id?: string
+  value: string
+}
+
+export interface CreateProductOptionDTO {
+  title: string
+}
+
+export interface CreateProductVariantOptionDTO {
+  value: string
+}
+
+export interface CreateProductVariantDTO {
+  title: string
+  sku?: string
+  barcode?: string
+  ean?: string
+  upc?: string
+  allow_backorder?: boolean
+  inventory_quantity?: number
+  manage_inventory?: boolean
+  hs_code?: string
+  origin_country?: string
+  mid_code?: string
+  material?: string
+  weight?: number
+  length?: number
+  height?: number
+  width?: number
+  options?: CreateProductVariantOptionDTO[]
+  metadata?: Record<string, unknown>
+}
+
+export interface CreateProductDTO {
+  title: string
+  subtitle?: string
+  description?: string
+  is_giftcard?: boolean
+  discountable?: boolean
+  images?: string[] | { id?: string; url: string }[]
+  thumbnail?: string
+  handle?: string
+  status?: ProductStatus
+  type?: CreateProductTypeDTO
+  type_id?: string
+  collection_id?: string
+  tags?: CreateProductTagDTO[]
+  // sales_channel
+  categories?: { id: string }[]
+  options?: CreateProductOptionDTO[]
+  variants?: CreateProductVariantDTO[]
+  width?: number
+  height?: number
+  length?: number
+  weight?: number
+  origin_country?: string
+  hs_code?: string
+  material?: string
+  mid_code?: string
+  metadata?: Record<string, unknown>
+}
+
+export interface CreateProductOnlyDTO {
+  title: string
+  subtitle?: string
+  description?: string
+  is_giftcard?: boolean
+  discountable?: boolean
+  images?: { id?: string; url: string }[]
+  thumbnail?: string
+  handle?: string
+  status?: ProductStatus
+  collection_id?: string
+  width?: number
+  height?: number
+  length?: number
+  weight?: number
+  origin_country?: string
+  hs_code?: string
+  material?: string
+  mid_code?: string
+  metadata?: Record<string, unknown>
+  tags?: { id: string }[]
+  categories?: { id: string }[]
+  type_id?: string
+}
+
+export interface CreateProductVariantOnlyDTO {
+  title: string
+  sku?: string
+  barcode?: string
+  ean?: string
+  upc?: string
+  allow_backorder?: boolean
+  inventory_quantity?: number
+  manage_inventory?: boolean
+  hs_code?: string
+  origin_country?: string
+  mid_code?: string
+  material?: string
+  weight?: number
+  length?: number
+  height?: number
+  width?: number
+  options?: (CreateProductVariantOptionDTO & { option: any })[]
+  metadata?: Record<string, unknown>
+}
+
+export interface CreateProductOptionOnlyDTO {
+  product: { id: string }
+  title: string
 }